+++ backend/src/main/java/com/DXC/iotbackend/payload/SignInRequest.java
@@ -1,0 +1,16 @@
+//package com.DXC.iotbackend.payload;
+//
+//import lombok.Data;
+//import jakarta.validation.constraints.Email;
+//import jakarta.validation.constraints.NotBlank;
+//
+//@Data
+//public class SignInRequest {
+//
+//    @Email(message = "Invalid email format")
+//    @NotBlank(message = "Email is required")
+//    private String email;
+//
+//    @NotBlank(message = "Password is required")
+//    private String password;
+//}
