+++ backend/src/main/java/com/DXC/iotbackend/payload/SignUpRequest.java
@@ -1,0 +1,31 @@
+//package com.DXC.iotbackend.payload;
+//
+//import jakarta.validation.constraints.Pattern;
+//import lombok.Data;
+//import jakarta.validation.constraints.Email;
+//import jakarta.validation.constraints.NotBlank;
+//import jakarta.validation.constraints.Size;
+//
+//import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+//
+//@JsonIgnoreProperties(ignoreUnknown = false)
+//@Data
+//public class SignUpRequest {
+//    @NotBlank(message = "First name is required")
+//    private String firstName;
+//
+//    @NotBlank(message = "Last name is required")
+//    private String lastName;
+//
+//    @Email(regexp = "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$", message = "Invalid email format")
+//    @NotBlank(message = "Email is required")
+//    private String email;
+//
+//    @NotBlank(message = "Password is required")
+//    @Size(min = 6, message = "Password must be at least 6 characters")
+//    @Pattern(
+//            regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%^&*()_+={\\[\\]}-]).{6,}$",
+//            message = "Password must contain at least 1 uppercase, 1 lowercase, 1 digit, and 1 special character"
+//    )
+//    private String password;
+//}
