+++ backend/src/main/java/com/DXC/iotbackend/repository/UserRepository.java
@@ -1,5 +1,8 @@
 package com.DXC.iotbackend.repository;
 
+
+
+
 import com.DXC.iotbackend.model.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -9,7 +12,10 @@ public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     // Checks if an email already exists (used in registration)
     boolean existsByEmail(String email);
+    boolean existsByUsername(String username);
 
     // Fetches the user by email (used in login)
     Optional<UserEntity> findByEmail(String email);
+    Optional<UserEntity> findByUsername(String username);
+
 }
