--- backend/src/main/java/com/dxc/iotbackend/SecurityConfig.java
@@ -1,87 +1,0 @@
-package com.DXC.iotbackend;
-
-
-import com.DXC.iotbackend.JwtFilter;
-import com.DXC.iotbackend.RateLimitingFilter;
-import com.DXC.iotbackend.UserDetailsImpl;
-import com.DXC.iotbackend.model.UserEntity;
-import com.DXC.iotbackend.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-@Configuration
-@EnableMethodSecurity
-
-public class SecurityConfig {
-
-    private final JwtFilter jwtFilter;
-    private final RateLimitingFilter rateLimitingFilter;
-
-    @Autowired
-    public SecurityConfig(JwtFilter jwtFilter, RateLimitingFilter rateLimitingFilter) {
-        this.jwtFilter = jwtFilter;
-        this.rateLimitingFilter = rateLimitingFilter;
-    }
-
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        System.out.println(" SecurityFilterChain setup completed");
-        return http
-                .csrf(csrf -> csrf.disable())
-                .cors(cors -> {})
-                .authorizeHttpRequests(auth -> auth
-                                .requestMatchers("/api/login", "/api/register").permitAll()
-                                .requestMatchers("/api/admin/**").hasRole("ADMIN")
-                                .requestMatchers("/api/user/**").hasAnyRole("USER", "ADMIN")
-                                .anyRequest().denyAll() // default deny
-                        //.anyRequest().authenticated()
-                )
-                .addFilterBefore(rateLimitingFilter, UsernamePasswordAuthenticationFilter.class) // Rate limiting first
-                .addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class) //JWT authentication
-                .build();
-
-
-    }
-
-    @Bean
-    public UserDetailsService userDetailsService(UserRepository userRepository) {
-        return email -> {
-            UserEntity user = userRepository.findByEmail(email)
-                    .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-            return new UserDetailsImpl(user.getUsername(), user.getRole());
-        };
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authConfig) throws Exception {
-        return authConfig.getAuthenticationManager();
-    }
-
-
-    @Bean
-    public CorsConfigurationSource corsConfigurationSource() {
-        CorsConfiguration config = new CorsConfiguration();
-        config.addAllowedOriginPattern("*");
-        config.addAllowedHeader("*"); // allow all headers including Authorization
-        config.addAllowedMethod("*"); // GET, POST, etc.
-
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", config);
-        return source;
-    }
-
-}
