--- backend/src/main/java/com/dxc/iotbackend/repository/UserRepository.java
@@ -1,21 +1,0 @@
-package com.DXC.iotbackend.repository;
-
-
-
-
-import com.DXC.iotbackend.model.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
-    // Checks if an email already exists (used in registration)
-    boolean existsByEmail(String email);
-    boolean existsByUsername(String username);
-
-    // Fetches the user by email (used in login)
-    Optional<UserEntity> findByEmail(String email);
-    Optional<UserEntity> findByUsername(String username);
-
-}
